홈 화면 컴포넌트 분리 및 거주 리뷰 섹션 추가
